+++ tools/explorerEval/tool.js
@@ -110,6 +110,7 @@
         const sumElem = $('tr.sum', container);
         for (const newRow of newRows) {
           const uci = newRow.uci;
+          if (!uci) continue;
           const move = co.parseUci(uci);
           const san = co.san.makeSan(ch, move);
           if ($('td', container).filter((i, e) => $(e).text() == san).length) continue; //castling can be identified by multiple ucis (i.e. e1g1, e1h1)